refactored some functions to use as imports/ moved them to src
