+++ hmtl/index.html
@@ -17,6 +17,6 @@
         </nav>
     </header>
     <main id="main"></main>
-    <script src="../script/script.js"></script>
+    <script type="module" src="../script/script.js"></script>
 </body>
 </html>
