+++ script/script.js
@@ -2,15 +2,7 @@ const log = (logs) => console.log(logs);
 
 const LINK = `https://64ec7004f9b2b70f2bfa53e4.mockapi.io`
 
-const getData = async (endPoint) => {
-    try {
-        const response = await fetch(LINK + endPoint);
-        const data = await response.json();
-        return data;
-    } catch (err) {
-        return false;
-    }
-}
+import {buildContent} from "./src/sharedFuctions.js";
 
 const redirectToItemPage = (item) => {
     window.location.replace(`./item.html?itemId=${item.id}`);
@@ -55,39 +47,4 @@ const buildItemsWrapper = (data) => {
     return itemsWrapper;
 }
 
-const buildMessageWrapper = (message,status) => {
-    const messageWrapper = document.createElement(`div`);
-    messageWrapper.classList.add(`message`);
-    messageWrapper.innerText = message;
-    switch (status) {
-        case (`error`):
-            messageWrapper.classList.add(`error`);
-            break;
-        default:
-            messageWrapper.classList.add(`success`);
-    }
-    return messageWrapper;
-}
-
-const buildContent = (data) => {
-    if (data) {
-        return buildItemsWrapper(data);
-    } else {
-        return buildMessageWrapper(`Something went wrong while getting the data, please try again.`,`error`);
-    }
-}
-
-const displayContent = (content,wrapper) => {
-    wrapper.append(content);
-}
-
-const buildMain = async () => {
-    const main = document.getElementById(`main`);
-    const data = await getData(`/items`);
-    log(data);
-    const content = buildContent(data);
-    displayContent(content,main);
-}
-
-buildMain();
-
+buildContent(`${LINK}/items`,buildItemsWrapper,'main');
